pyproject.toml
[project]
name = "sdr-agent"
version = "0.1.0"
description = "SDR Agent - Multi-agent system for sales development research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "mcp>=0.9.0",
    "langsmith>=0.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "ddgs>=9.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0"
]

[tool.setuptools]
packages = ["agents", "core", "tools", "utils"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta" 